+++ packages/preset/src/prefixes.ts
@@ -79,6 +79,8 @@ export default [
   'radio-group-item',
   'resize',
   'rounded',
+  'scroll-area',
+  'scroll-bar',
   'select',
   'select-content',
   'select-group',