+++ uno.config.ts
@@ -45,6 +45,7 @@ export default defineConfig({
   configDeps: [
     '../packages/nuxt/src/preset.config.ts',
 
+    '../packages/preset/src/_shortcuts/scroll-area.ts',
     '../packages/preset/src/_shortcuts/sheet.ts',
     '../packages/preset/src/_shortcuts/toast.ts',
     '../packages/preset/src/_shortcuts/form.ts',